backend/pyproject.toml
[project]
name = "code-review-agent"
version = "0.1.0"
description = "AI-powered Self-Correction Agent for Automated Code Reviews"
authors = [
    {name = "AI Systems Architect"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "sqlalchemy[asyncio]>=2.0.29",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "langgraph>=0.0.38",
    "langchain-core>=0.1.45",
    "openai>=1.19.0",
    "anthropic>=0.23.0",
    "chromadb>=0.4.24",
    "celery>=5.3.6",
    "redis>=5.0.3",
    "structlog>=24.1.0",
    "prometheus-client>=0.20.0",
    "docker>=7.0.0",
    "websockets>=12.0",
    "sse-starlette>=2.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "C4", "SIM", "RUF"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
